Support phony input when recording mtime of cleaned node
This is a fix for issue #2405 where a target with a single phony
input would be considered dirty forever after being built if the
rule was setting `restat = 1` because the code would try to stat
a non-existing file to record the input mtime.

Instead use the mtime recorded in the Node instance (restating it
if necessary) to avoid this issue.